refactor with react-router to allow multiple pages
