+++ front-end/src/pages/TaskCreationPage/index.js
@@ -1,0 +1,1 @@
+export * from "./TaskCreation"
